+++ NEWS.md
@@ -1,6 +1,7 @@
 # DEVEL
 
 * Fixed: `as.data.table()` method for `featureGroups`: `normConcToTox` argument was ignored (not for `featureGroupsScreening`)
+* Various reporting fixes for screening results from suspect lists without formula/SMILES data.
 
 # patRoon 2.3.1